Initial commit
This is the package, (almost) as generated by https://appgenerator.yunohost.org/
(The only change is to add the DB dump command to the backup script.)
Most of the documentation things are still left for later.
The config.inc.php comes from a working install of MRBS as a my_webapp,
modified to work as a template.