Split config into base and customization
config.inc.php with the customization stuff
is for the admin to edit as they please, while
config-ynh.inc.php does the behind-the-scenes work to integrate MRBS
with YunoHost (DB, Auth).
config.inc.php loads config-ynh.inc.php with a `require`.
Making this separation should be useful, because it allows us to change
config-ynh.inc.php at will, without having to deal with the config
system to keep customization intact.

Sidenote, I decided to leave the timezone in config-ynh.inc.php since
changing it after the installation may cause difficulties.