Create timezone setting for upgrade script
Since the upgrade script templates out config-ynh.inc.php, it needs the
timezone variable to be set.
I decided to create a ynh_setting for that, since the point in including
it in config-ynh.inc.php is that it should not be changed without the
admin being aware of and dealing with the consequences, and they might
have changed the server timezone before the upgrade.